Tests/LibWeb/Text/expected/svg-relayout-self-size-change.txt

svg=0,0,300,200
after=0,200,80,20
PASS
